ProjectComplete

+++ CourseProject.Business/Services/AddressService.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using CourseProject.Business.Exceptions;
 using CourseProject.Business.Validators;
 using CourseProject.Common.DTOs;
 using CourseProject.Common.DTOs.Address;
@@ -34,7 +35,14 @@ public class AddressService : IAddressService
 
     public async Task DeleteAddressAsync(AddressDelete addressDelete)
     {
-        var entity = await AddressRepository.GetByIdAsync(addressDelete.Id);
+        var entity = await AddressRepository.GetByIdAsync(addressDelete.Id, (address)=> address.Employees);
+        
+        if (entity == null)
+            throw new AddressNotFoundException(addressDelete.Id);
+
+        if (entity.Employees.Count > 0)
+            throw new DependentEmployeesExistException(entity.Employees);
+        
         AddressRepository.Delete(entity);
         await AddressRepository.SaveChangesAsync();
     }
@@ -42,6 +50,10 @@ public class AddressService : IAddressService
     public async Task<AddressGet> GetAddressAsync(int id)
     {
         var entity = await AddressRepository.GetByIdAsync(id);
+
+        if (entity == null)
+            throw new AddressNotFoundException(id);
+
         return Mapper.Map<AddressGet>(entity);
     }
 
@@ -54,6 +66,11 @@ public class AddressService : IAddressService
     public async Task UpdateAddressAsync(AddressUpdate addressUpdate)
     {
         await AddressUpdateValidator.ValidateAndThrowAsync(addressUpdate);
+
+        var existingAddress = AddressRepository.GetByIdAsync(addressUpdate.Id);
+       
+        if (existingAddress == null)
+            throw new AddressNotFoundException(addressUpdate.Id);
 
         var entity = Mapper.Map<Address>(addressUpdate);
         AddressRepository.Update(entity);
